split out flake detection from utility to detect changed tests
